web2/requirements.txt
Flask==1.0.3
grpcio-tools==1.21.1
